requirements.txt
discord.py==2.0.1
numpy==1.22.3
pandas==1.4.2
python-dotenv==0.21.0
setuptools==63.2.0
youtube_dl==2021.12.17
# youtube_search_python==1.6.6
youtube_search==2.1.2
pynacl==1.5.0
pytest
pytest-asyncio
